+++ python/tskit/tables.py
@@ -2080,7 +2080,12 @@ class TableCollection:
 
     @property
     def indexes(self):
-        return TableCollectionIndexes(**self._ll_tables.indexes)
+        indexes = self._ll_tables.indexes
+        return TableCollectionIndexes(**indexes) if indexes is not None else None
+
+    @indexes.setter
+    def indexes(self, indexes):
+        self._ll_tables.indexes = indexes.asdict()
 
     @property
     def sequence_length(self):
@@ -2134,7 +2139,7 @@ class TableCollection:
         Note: the semantics of this method changed at tskit 0.1.0. Previously a
         map of table names to the tables themselves was returned.
         """
-        return {
+        ret = {
             "encoding_version": (1, 1),
             "sequence_length": self.sequence_length,
             "metadata_schema": str(self.metadata_schema),
@@ -2148,6 +2153,9 @@ class TableCollection:
             "populations": self.populations.asdict(),
             "provenances": self.provenances.asdict(),
         }
+        if self.indexes is not None:
+            ret["indexes"] = self.indexes.asdict()
+        return ret
 
     @property
     def name_map(self):
@@ -2278,7 +2286,6 @@ class TableCollection:
             tables.metadata = tables.metadata_schema.decode_row(tables_dict["metadata"])
         except KeyError:
             pass
-
         tables.individuals.set_columns(**tables_dict["individuals"])
         tables.nodes.set_columns(**tables_dict["nodes"])
         tables.edges.set_columns(**tables_dict["edges"])
@@ -2287,6 +2294,12 @@ class TableCollection:
         tables.mutations.set_columns(**tables_dict["mutations"])
         tables.populations.set_columns(**tables_dict["populations"])
         tables.provenances.set_columns(**tables_dict["provenances"])
+
+        # Indexes must be last as other wise the check for their consistency will fail
+        try:
+            tables.indexes = TableCollectionIndexes(**tables_dict["indexes"])
+        except KeyError:
+            pass
         return tables
 
     def copy(self):